refactor(backend): align code with AGENTS.md conventions
- Add logger.exception in security.py JWTError handler
- Add deleted_at filter to user query in auth dependency
- Remove duplicate EmailAlreadyRegisteredException class
- Fix lazy imports in auth routes and dependencies
- Add _active_filter helper to AuditLogRepository
- Change Mapped[float] to Mapped[Decimal] for all money fields
- Remove float casts in expense repository analytics
- Change debt repository update_remaining_amount to Decimal
- Clean up float branch in expense_service _serialize_value
- Create InstallmentRepository interface and implementation
- Fix InstallmentService to use repository instead of raw session
- Remove double-commit by letting UoW manage transactions
- Add .opencode to .gitignore

+++ backend/app/domains/repositories/debt.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from decimal import Decimal
 from typing import List, Optional
 
 from app.models import Debt
@@ -26,5 +27,5 @@ class DebtRepository(ABC):
         pass
 
     @abstractmethod
-    async def update_remaining_amount(self, debt_id: str, new_amount: float) -> Debt:
+    async def update_remaining_amount(self, debt_id: str, new_amount: Decimal) -> Debt:
         pass
